﻿[DotNetRCAndUtils:823] Fixed issue where the query edit box would lose focus when the query dirty status changed.
[DotNetRCAndUtils:824]
Fixed issue that allowed an invalid query to be re-evaluated before the execute button was pressed. Also made a change so that if a query fails to evaluate when first opened (whether it is because the query is bad, or the DB schema is no longer in sync with the query), the tab for the query will now open so that a user can correct it.

[DotNetRCAndUtils:825]
Added ability for the SQLCDBEditor to call a batch file or VB script on save if the script has the same name as the database with a “.bat” or “.vbs” extension. (e.g. “Ordermapper.sdf.bat”)

Review pending…

[git-vault-id] Extract$/Engineering@19857/21454