﻿[DotNetRCAndUtils:829] To solve the case where the user could end up clicking on an un-intended cell because the data status label got inserted and pushed the grid down during a click, the panel that displays the data status label is now always visible whether or not the label itself is.
[DataEntry:729]
The code that allowed the DataEntry UI to ensure it had the latest copy of the DE SQLC database with each new document opened was broken (apparently with the addition of multi-DEP support. The DataEntry UI should once again always be grabbing the latest copy of the database with each new file loaded.

Added retries for all places that copy to or from the master database copy where there weren’t already retries.

Fixed yet another case where the valid/invalid state of a table was not properly updated.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@20008/21605