﻿One missing piece to the new system for handling AKAs that uses the "Data Query" rule object and "Add AKAs" SQLCDBEditor plugin is that data queries need to be able to output spatial coordinates to the database for use by the Add AKAs plugin. I made these updates to queries to make this possible:
[DataEntry:1126]
Added a SpatialField output that can convert the overall bounds of a spatial or attribute result to a text value representing any of the following fields:
Page, Left, Top, Right, Bottom, StartX, StartY, EndX, EndY, Height

[DataEntry:1127]
Added an AbortIfEmpty attribute that causes the parent query node to abort evaluation and return blank if a query node with this attribute is blank. This makes it less cumbersome to deal with compound queries where a nested query that ordinarily would return a list instead returns nothing.

In the process of creating new automated tests, found and fixed an error that could cause a list of query results to be truncated.

Created new automated tests for the above features.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@20753/22358