﻿Checking in a first draft of the UITC DEP.
The DEP has a new level in the attribute hierarchy: “Panel” which contains the dates/times and lab data.

Included in this are additions in tweaks to queries such as the flag validation field being able to handle commas and > or < in the value field.

The laboratory fields need some work to conform to the new schema yet, but I need to move on to other things for the moment.

[git-vault-id] Extract$/Engineering@20792/22397