﻿[DataEntry:1141] Fixed case where query results were unnecessarily being marked dirty which caused UI values to be updated which, in turn, caused more queries to be evaluated. In some cases, such as when deleting a panel from the UITC DEP, this could lead to exponentially more query evaluations than necessary.
[DataEntry:1131, 1144]
Fixed a situation in which validation queries could trigger a “collection was modified” exception. The above issue contributed to this issue.

These fixes also solve a case where, at least in the build 240 code, when typing value in a table cell with an auto-complete list, the value could be cleared after typing the last character of an auto-complete value. Specifically, in Demo_FlexIndex Example01 in the zip code field after typing the “9” in 63179, the field would be cleared.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@21333/22944