﻿[DataEntry:1146] Fixed case where some auto-update queries would not fire.
[DataEntry:1130]
Added property too data entry controls to specify whether or not they should display highlights for descendant attributes that are in other controls. The property defaults to true for tables and false for all other controls.

[DataEntry:495]
Added ability for swiping to be disabled on a row or column basis in tables.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@21346/22957