﻿[DataEntry:1109, 621] Fixed some auto-complete behavior issue in the new row of tables.
Fixed a problem with a recent fix for excessive auto-update triggering that would prevent data from being pasted into a table cell already in edit mode.

Fixed a problem where after pasting data into a cell in the new row, that cell would be selected, but the row selection would be in the new row.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@21361/22972