﻿[FlexIDSCore:4996] Fixed a couple of situations that could lead to extremely long processing time when trying to use the Find and Redact tool to add redactions for terms with many hits on documents with many pages.
On my dev machine using a test document 1400 pages long and a search term that garnered 750 hits, it appeared it would have taken hours to complete. I can now add all redactions in about 45 seconds; I’m sure there is room for improvement, but I believe this is sufficient for now.

[FlexIDSCore:5167]
When the Find and Redact window is checking if hits are redacted before applying a new redaction, it will no longer consider turned off redactions; in the case that one of the hits was redacted but turned off, a new redaction will now be added.

[FlexIDSCore:5166]
Fixed errors that could occur in the ID Shield verification UI if the “Discard Changes” option were used after the word highlight tool had previously been activated on a document but while it was not still active.

[LegacyRCAndUtils:6327]
Fixed possibility of the ActionStatisticsDelta growing unmanageably large if processing is being done solely by service threads with no FAM or DBAdmin processes checking stats from time to time.

Reviewed by:
William (0/3)

[git-vault-id] Extract$/Engineering@21780/23394