﻿[LegacyRCAndUtils:6337] Restructured path tag/function evaluation so that path tags are now expanded within the context of individual path function parameters instead of universally across the all text before function evaluation. This means the expanded contents will not affect the parsing of the overall parsing of functions and parameters. For example, <SourceDocName> values can now contain commas and parenthesis.
[LegacyRCAndUtils:6331]
It turns our LegacyRCAndUtils:5710 had already been implemented (and apparently forgotten) to allow argument tokens other than “,” to be specified by including the new token in brackets before the opened parenthesis. But I have improved on this fix so that it will now apply universally to all path tag functions rather than for just the functions that explicitly support it. For example:
$Replace{|}(<SourceDocName>|,|_) can be used to replace all commas in the SDN with underscores. Note that because of the fix for LegacyRCAndUtils:6337, alternate token specification needs to occur only if the literal text of a function parameter contains commas.

The fix for [LegacyRCAndUtils:6327] introduced an issue where the FAM would hang when stopping for up to 45 seconds if stopped shortly after starting. I have fixed this issue.

Review pending…

[git-vault-id] Extract$/Engineering@21993/23607