﻿[LegacyRCAndUtils:6363] The path tag/function rework revisited. The original goal of expanding path tags only within the context of a particular path tag argument has now been implemented such that it does not cause a failure to parse cases where path tag functions that have multiple parameters are nested within another path tag function.
[DataEntry:1125, 1158]
The data query output handler for Fresno was made much more cumbersome that it would otherwise need to be by not being able to select the name of an attribute. Since it was pretty simple to implement, I added an AttributeField attribute that allows selection of attribute names and types and added support for full AF query syntax in attribute query nodes. Now the Fresno AddToRecentlyUsedDB needs just 3 queries instead of 6 sets of 3 (1 set for each attribute for which used values are being collected).

Added automated tests for the query enhancements.

Reviewed by:
William (0/1)
[git-vault-id] Extract$/Engineering@22086/23700