﻿[DotNetRCAndUtils:806] Fixed case where Arg1 of the FTPEventHistory table was not correct when queuing files from subdirectories of the root download folder.
[DotNetRCAndUtils:844]
Added licensing for DataEntryPrompt utility.

[DotNetRCAndUtils:847]
First field of the DataEntryPrompt utility now gets focus when form is first displayed.

[DotNetRCAndUtils:849]
It had been possible for the rasterization process in the rasterize PDF task to crash without the task knowing about it. This, in turn could cause the source pdf to be inappropriately deleted or renamed in the database. The rasterize PDF task is now able to detect a rasterization process that crashed.

[DotNetRCAndUtils:851]
If Rasterize PDF task fails with one rasterization method, the other will be attempted automatically.

Reviewed by:
William (0/1)

[git-vault-id] Extract$/Engineering@22107/23721