﻿[DotNetRCAndUtils:835] Added code to protect against the circumstance that an FPS file configured to keep running as files are added to the queue could fail to re-spawn when running under the FAM service.
Also added a couple application traces for the circumstance that a FAM is set to run only a specified number of files (such as via the FAM service), but has supplying enabled. One warns when processing starts that the instance may not be able to stop or restart and the second warns at such a time that the specified number of files have been processed and it is indeed unable to do so.

[DotNetRCAndUtils:836]
If the FTP file supplier fails, it will now automatically restart itself if a polling mode is enabled.

[git-vault-id] Extract$/Engineering@22205/23820