﻿[FlexIDSCore:5186] Fixed a situation that would lead to “Task ID not found in map” app traces being logged.  It appears that there is at least one more situation that could lead to this, but at a minimum, the number of these log messages should be significantly reduced.
[FlexIDSCore:5194]
Made a change that  I think has a good chance at solving the situation where an exception occurs when right-clicking on a file in the completed list of the progress log FAM tab.

[FlexIDSCore:5193]
Made some changes that I think have a chance at solving the crash that is occurring in the Extract backup FAM. If the crash is not solved, this change should generate more data to help solve it.

Reviewed by:
William (1/0)

[git-vault-id] Extract$/Engineering@22612/24228