﻿[DataEntry:1176] Fixed situation that allowed the DE host’s ItemSelectionChanged to be raised before the statistics related to the newly selected attributes were updated. This could prevent the and I suspect may have been involved with some cases of DataEntry:867 where the next invalid item button was not properly disabled when the last invalid data item was corrected.
[DataEntry:1177]
Fixed situation where un-viewable attributes could be assigned highlights when “accept spatial info” is used.

Also made a tweak so that when using the undo operation to undo an “accept spatial info” for a table row, the entire accept is now done at once instead of one highlight at a time.

Review pending…

[git-vault-id] Extract$/Engineering@22914/24530