﻿[DataEntry:1178] The show all highlights option no longer show highlights of attributes that are mapped to hidden controls.
[DataEntry:1179]
After accepting spatial info, the state of the accept spatial info item now properly reverts back to disabled.

[DataEntry:1180]
Also fixed an issue where a data query SpatialMode of “Only” would allow text to be returned if there was no spatial info to begin with.

Also fixed a tab order issue that was introduced with the last checkin.

Reviewed by:
William(0/0)

[git-vault-id] Extract$/Engineering@22924/24540