﻿[LegacyRCAndUtils:6327] In my previous note where I said "If this was doing any kind of processing at all (even just copying a file), we wouldn't run into this"... I realized that test IS copying (and deleting) files.
I decided to go ahead and drop the interval that the stats maintainence thread checks if the delta table needs flushing to 10s from 60s. This should be a safe change and even at 10s, it is still putting significantly less stress on the DB than the status bar of a running FAM will (which hits the DB every 2 seconds with a larger query).

I'm re-closing this SCR since in seems to me very unlikely that there will ever be a real setup in which FAMProcesses are re-spawning in < 10s.

Review pending…

[git-vault-id] Extract$/Engineering@22950/24566