﻿[DataEntry:1186] Fixed issue where data queries could cause data to be restored to a different state than originally existed after preforming an undo operation.
[DataEntry:1188]
Fixed situation in tables where data was refreshed in the middle of trying to apply new attribute data. This refresh could cause loss of spatial information and for auto-update queries to be executed pre-maturely such that the original data would never be properly applied (such as when swiping in a new row or copying one row over top of another row). I suspect this may fix related exceptions that cannot be consistently reproduced, but I’m not sure so I will be keeping this SCR open for the time being.

[DataEntry:1199]
A wait cursor is now displayed when processing swipes with the word highlight tool.

Review pending…

[git-vault-id] Extract$/Engineering@23019/24635