﻿[DataEntry:1102, 1201, 1206] Fixed smart tag behavior  in table cells and controls using auto-complete including that smart tags should no longer be at risk of generating an exception.
[DataEntry:1200]
Fixed issue that could cause a zoom into the selection despite being set to “Zoom out if necessary” and the last manual zoom level not having been zoomed in.

[DotNetRCAndUtils:819, 882, 885]
To prevent getting data into a bad state when data can’t be merged due to the lack of a primary key on a table, tables will now need a primary key to be editable. (not just a unique constraint, which was the previous requirement). The status label will now indicate when a table is not editable.

Also, tweaked the query to collect the original test name for the purposes of the InsertAKAs system so that it will work even when the name is obtained via a swipe of only the name field (rather than via a swipe of the whole row).

[git-vault-id] Extract$/Engineering@23084/24700