﻿[DataEntry:1208] Fixed case in which use of a smart tag in a text box could end up modifying the last selected table cell.
[DataEntry:1209]
Swept an apparently not straight forward exception related to smart tags under the rug by making it a log instead. Worst case scenario, the smart tag doesn’t get applied (and we have something logged), but per the exception, it seems it wasn’t intended that a smart tag should be added at the time of this exception.

[FlexIDSCore:6391]
Fixed problem in SQL to export the top (rather than random) files. Could not reproduce issue with use of the randomized selection. I’m going off the theory that in Wayne’s testing, the problem from the top query left a DB setting in the wrong state.

[FlexIDSCore:5216]
When the IDShieldTester is configured to test multiple folders, the file lists will once again contain files from all folders, not just the last.

Review pending…

[git-vault-id] Extract$/Engineering@23130/24748