﻿[DataEntry:1230] Fixed case where cursor tools could appear to get “stuck” in shift/ctrl states after using those modifier keys in the DEP.
[DataEntry:1232]
Fixed cases where “Failed to apply updated value” exceptions could be displayed for documents that had been run through rules under a different filename.

[DataEnty:1188]
Implemented code to allow for clipboard copy retries for table rows or multiple cells that are copied to the clipboard as well as the use of an internal backup of the clipboard data that can be used for up to 15 seconds after the copy in case the copy never succeeds.

Review pending…

[git-vault-id] Extract$/Engineering@23267/24885