﻿[DotNetRCAndUtils:961] Shortcut keys now work if either the image or thumbnail panes have focus.
[DotNetRCAndUtils:962]
The output filename browse button is now enabled/disabled correctly based on selection.

[DotNetRCAndUtils:965]
Clicking on page thumbnail controls now only selects unless the control key is the only modifier key.

[DotNetRCAndUtils:966]
Shift-key selection behavior is now consistent with Window’s standards in that the selection can be reduced as well as expanded.

[DotNetRCAndUtils:967]
Toggle document separator option is now available only when a single page is selected.

[DotNetRCAndUtils:968]
It is no longer required for a control to be previously selected before it can be dragged.

Also:
- Fixed problem where the drop indicator would not appear if the mouse were dragged directly between two page controls.
- Dramatically reduced the number of GDI handles the application uses, particularly when a large number of pages are loaded.

Review pending…

[git-vault-id] Extract$/Engineering@24032/35649