﻿[DotNetRCAndUtils:993] Fixed bad text in file selection box when connecting to a DB with an old schema.
[DotNetRCAndUtils:995]
Made a change to the random file selection query so that it is not hampered by an issue with pre-compiling that seems to occur when the query is run from the FAMFileInspector.

[DotNetRCAndUtils:996]
Fixed handling of a failed login to a database after using the file inspector’s logout option.

[DotNetRCAndUtils:998]
Fixed exception that would occur if a regex search returned an empty result.

[DotNetRCAndUtils:999]
Fixed ability to sort by match count when a data (attribute) search has been performed.

Also:
- When the file selection is changed, previous search results are now reset.

Reviewed by:
William (1/1)

[git-vault-id] Extract$/Engineering@24412/46224