﻿[DotNetRCAndUtils:1016, 1017] Fixed issue that would cause any attempted search operation attempted after a failed search attempt to fail with the same error. Also, search attempts on individual files will be allowed to fail without failing the entire search. (An aggregate exception will be shown on completion for all errors).
[FlexIDSCore:5244]
As a temporary stop-gap measure to try to prevent the FAM hang on stop issue, a new TryGetStats FileProcessingDB method has been added which will be called on stop in the FAM in place of the normal GetStats call. The TryGetStats method will not attempt a DB lock (then hangs appear to be while waiting for a lock in GetStats). If unable to get stats without locking the DB, an app trace will be logged and the stats at the bottom of the FAM may be out of date.

Review pending…

[git-vault-id] Extract$/Engineering@24471/46283