﻿[DotNetRCAndUtils:1028] The Arguments column of the LaunchApp table can now be null.
[DotNetRCAndUtils:1029]
Fixed a number of different scenarios that could cause crashes:
- Closing FFI while a search or non-blocking app launch operation was still running.
- Closing FFI while exception related to failures in a search or app launch operation was displayed.
- Multiple non-blocking launch app operations started in quick succession.

If a non-blocking app launch operation is running at the time of close, a prompt will now be displayed asking whether the user wants to stop the operation before the next file is processed and close or whether to cancel the close and keep the operation running. If so, a non-cancelable modal message box will be displayed until the current file is done processing, then the form will close. The process that is operating on the current file will not be forcefully killed.

Review pending…

[git-vault-id] Extract$/Engineering@24533/46354