﻿[FlexIDSCore:5293] Implemented a system to address memory usage of unmanaged COM objects referenced from managed code.
We have long had issues with the garbage collector being lazy when it comes to releasing unmanaged COM objects; the reason is that the garbage collector was unaware of the memory associated with the unmanaged COM object references it was holding. Via the implementation of a new IManageableMemory interface implemented by certain c++ COM classes and a IMemoryManager implemented on the .Net side, most unmanaged COM objects referenced by .Net rule objects will now report their memory usage to the garbage collector to allow it to manage memory appropriately.

The end result is that memory usage when running rules that make use of .Net rule objects will behave much better, particularly in cases that involve use of the input finder.

Review pending…

[git-vault-id] Extract$/Engineering@25455/47339