﻿[LegacyRCAndUtils:6460] Added code to ignore exceptions being caused by the new memory management system for use of unmanaged memory objects in .Net COM objects (FlexIDSCore:5293).  It seems that when a process is ending, the calls attempted by the unmanaged memory objects to the managed MemoryManager object to try to release their memory they used fail because "The process terminated unexpectedly." I am unclear on why the .Net MemoryManager instance would have apparently gone away before the object referencing it has been released.
In any case, this is an innocuous exception. There is no need to be concerned with reporting correct unmanaged memory usage to the garbage collector if the process has gone away. I have therefore added code to ignore this exception when trying to report memory usage as long as the underlying error is 0x8007042b (The process terminated unexpectedly).

I have left the SCR open, but un-tagged it for 9.6 (in the long run, the underlying cause of the exceptions should be fixed rather than ignoring the exceptions)

Review pending…

[git-vault-id] Extract$/Engineering@26500/68383