﻿[DataEntry:1242] Re-enabled F6 shortcut to zoom to/from current selection. This had been disabled just prior to the 9.1 release in order not to push back the release further with a fix for a related exception. As part of testing this the feature I found and fixed the following minor issues:
[DataEntry:1275]
F6 will now account for the tooltip when zooming to a field that is not on the currently displayed page in the image viewer.

[DataEntry:1276]
The “Zoom out when necessary” auto-zoom setting will no longer inappropriately zoom in in certain cases.

[DataEntry:1277]
When the auto-zoom settings are not allowing for any context, auto-zoom (either via F6 or the “automatic” setting) will no longer zoom in a little too far.

Reviewed by:
William (1/0)
Follow-up review pending…

[git-vault-id] Extract$/Engineering@26639/68555