﻿[LegacyRCAndUtils:6450] Fixed issue where USBLicenseKeyManager utility was not aware of email settings changes made in the current session.
[LegacyRCAndUtils:6451]
If USBLicenseKeyManager is opened on XP or Server 2003 where we cannot for elevated privileges, the utility will now throw exceptions if settings are changed which cannot be stored (rather than silently ignoring the problem). Also, when opened in non-admin mode, the utility will now correctly display alert level and multiple values (which it had not been previously).

[LegacyRCAndUtils:6471]
Put the /RemovePages option for the ImageFormatConverter under a special license.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@26700/68616