﻿[DataEntry:1187] Fixed some quirky behavior with the F6 to zoom in/out including an issue where hitting F6 while zoomed to the selection could result in a view area that doesn’t include the currently selected field.
- Fixed several data query issues introduced with the 9.1 enhancements that necessitated ugly hacks in a number of situations. Ensured coverage for these issues in the NUnit tests:

[DataEntry:1225]
With the distinct selection mode, now the overall result will be combined as a single result even if there was only one result in the distinct node itself.

[DataEntry:1238]
Relative attribute queries now work even when the base attribute is not in the declaration or in a distinct node.

[DataEntry:1263]
Newlines are not interpreted literally rather than as result divider in SQL and expression nodes so that spaces to not needs to be manually added to the end of each line in a SQL query or expression to get it to evaluate correctly.

TESTING NOTE:
The specific problems are tested as part of the DataEntry NUnit tests. What needs manual attention is testing DEPs to see that the queries are still behaving as expected.

Review pending…

[git-vault-id] Extract$/Engineering@26775/68691