﻿[FlexIDSCore:5362] Since the core algorithm in the MergeAttributes OH  (as well as the Compare ID Shield data files contents condition) is order of magnitude (O^2), performance can get out of hand on documents with a large number of attributes (hundreds+) despite the fact that critical data used in the algorithm has been cached.  To prevent performance issues, the algorithm has been modified so that it processes each page individually which should prevent the algorithm from even having to deal with enough attributes that performance becomes an issue.
Fixed an issue with the LabDE rates report so that counts (docs, pages, orders, tests), come from the last verification session in which counts were saved instead of the last verification session overall. This prevents the possibility of documents, tests, etc from going uncounted if they are opened up post-save, but not subsequently re-saved.

Review pending…

[git-vault-id] Extract$/Engineering@26801/68717