﻿RC.Net\DataEntry RC.Net\Utilities
[DataEntry:1283, 1284, 1289]
Fixed a problem with the data entry query system that could cause queries (particularly complex queries) to needlessly and excessively update values before the changes that trigger the updates have been fully processed. The end result was that the same query at times ended up executing hundreds of times in response to a single attribute value changing. Not only did this severely affect performance of queries reacting to changes in the UI, but it could cause attribute queries to be resolved inconsistently since an attribute might have one value for one part of the query and a different value for another part.

Reviewed by:
William (0/2)

[git-vault-id] Extract$/Engineering@26924/68877