﻿Rather than wait until Arvind gets back to decide whether 9.6 needs to be held up for a fix for FlexIDSCore:5373, I have checked-in a work-around for FlexIDSCore:5373 as well as a pair of fixes that dramatically improve the speed of the “find matching numbers” logic on large documents in order to make a build.
If it is decided that these changes should not be in 9.6, the changes will be rolled back.

[FlexIDSCore:5373]
Disabled the new memory management system for the SpatialPageInfo class in order to work around memory and speed issues it appears to create. The SCR is being kept open since I do not understand the reason for the problems. (The SpatialPageInfo class is small enough that it probably does not need to use the new memory management system anyway).

Fixed an issue in SelectPageRegion that caused each page of a document to be read into a SpatialString even to select just one of the pages.

Added caching to ExtractOCRText so that it does not need to re-read the original OCR data from the uss file every time it is called. This was particularly inefficient as part of the “findmatchingnumbers” ruleset.

Review pending…

[git-vault-id] Extract$/Engineering@27209/69168