﻿[FlexIDSCore:5379] Enhanced format string syntax to support: - New variable term to allow expansion of a variable attribute via a nested format string - New variable term to allow specification of which value(s) to select when a variable query matches multiple attributes - New variable term to allow specification of a delimiter for the case that multiple values are selected. - All characters in format strings can now be escaped with a backslash to prevent them from being interpreted as part of the format string syntax.
[FlexIDSCore:5381]
Added support for back-referencing attributes passed into a rule-set from a parent ruleset or the execute rules task via a new syntax that can be used in the regexes of most rule objects. The syntax is similar to the fuzzy syntax, except the 4th char is a %:
(?~%FormatString)
Note that the % is in addition to the % char that will frequently come at the beginning of format strings. For example, to insert the first SSN value passed into the ruleset:
(?~%%*@SSN%)
Back-referencing format strings will be evaluated before fuzzy expressions. Both can be used in the same expressions, but if nested, the fuzzy expression needs to be defined within the back-referencing format string using escape chars if necessary. For example, to perform a fuzzy search for all unique SSN attributes passed in:
(?~%(?~\<\>%*@SSN<>uniq:\) | (?~\<\>%\)

Since it was not straight-forward to implement, support for back-referencing format strings was not added to these rule objects:
DocumentClassifier
EntityNameDataScorer
MERSHandler
RegularExpressionInputValidator

It was added to these rule objects:
SelectPageRegion
BlockFinder
BoxFinder
LocateImageRegion
MicrFinder
RegexRule
REMPFinder
ValueAfterClue
ValueBeforeClue
AdvancedReplaceString
RemoveCharacters
ReplaceStrings
PersonNameSplitter

Also, fixed errors in image view nunit tests and added new nunit project that allows only the automated nunit tests to be run (ignores interactive).

Reviewed by:
William (1/5)
[git-vault-id] Extract$/Engineering@27381/69343