﻿Part of the previous checkin was an enhancement of SpatialString to index the first character of each page—I assumed this would increase performance on large documents since the code would not have to iterate each letter up until it got to a requested page. In practice, I was surprised to find this change appears to have made performance about 5% worse on the large document test I am using.  I have removed this enhancement.
[FlexIDSCore:4753]
This SCR had supposedly been fixed—it was to allow document classification to be skipped when the classifier had already run. However, this was only the case when a document was successfully classified. If a classifier had run, but had failed to classify the document, classifiers in nested rules would continue to unnecessarily re-run. I have fixed this situation.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@27666/69636