﻿Further enhanced the new format string syntax to prevent having to use rule object to create temporary sub-attributes for the initial use case in the findmatchingnumbers ruleset.
Specifically, added the ability to replace the AFQuery of a variable with a regular expression prefixed by an equal sign.

If the regex is used in place of an AF query, rather than search existing sub-attributes, an attribute will be created and returned for each match of the regex in the source attribute's value.

Example:
%*@SSN<%=\\d<>all:.?%>uniq: | %

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@27688/69658