﻿[FlexIDSCore:5387] Enhanced the persistence of attribute vectors into VOA files so that the page info containing the data for all pages of a document is not repeated for each attribute in the VOA file. On large document with lots of attributes, this dramatically reduces the size of the VOA file created and, more importantly, dramatically lowers the memory usage of any process that loads the saved VOA file.
NOTE: This change will mean that older versions of the software will not be able to load VOA files saved from 9.7. (Though 9.7 will obviously continue to be able to load older versions of VOA files).

Also cleaned up a couple of warnings from previous check-ins that were not caught.

Reviewed by:
William(0/4)

[git-vault-id] Extract$/Engineering@27744/69714