﻿$/Engineering/ProductDevelopment/Common
LegacyRCAndUtils #5912
Updated RegisterAll.bat so that the search for the files containing the dll names that need to be registered only uses the file name since the current folder is already the CommonComponents folder.  This does mean the .rl and .nl files must not have spaces in their names.

William

[git-vault-id] Extract$/Engineering@27758/69728