﻿$/Engineering/ProductDevelopment/AFIntegrations/SharePoint/Redaction/Core/Code
Checked in to main branch

Reviewed by Steve(0/1)

William

From: William Parr [mailto:william@extractsystems.com]
Sent: Tuesday, January 28, 2014 7:51 AM
To: ag@extractsystems.com; steve_kurth@extractsystems.com; wayne_lenius@extractsystems.com
Subject: Sharepoint changes checked in - to fix problems in Production for Fl Supreme Court

$/Branches/FlexIndex/9.7.1/Engineering/ProductDevelopment/AFIntegrations/SharePoint/Redaction/Core/Code

Fixed problem where .Processed files not being cleaned up if there is no .redacted file
In the case when a file was processed and there was no items to redact found the .processed file to trigger the change in status in SharePoint did not get deleted.

Fixed problem where a document was deleted from SharePoint while it was being processed in the working folder.
This would cause the files in the working folder to never be cleaned up and the trigger file would be seen every time the timer job was ran causing an exception to be logged.

Fixed problem where RedactNow button and Verify button would display an error if there are more files in the Document Library than the List View Threshold
A dialog would be displayed showing the message “The attempted operation is prohibited because it exceeds the list view threshold enforced by the administrator.”

These changes were checked into the 9.7.1 branch and I have made a build 9.7.1.1
Still need to check in to the main branch

Review pending

William

[git-vault-id] Extract$/Engineering@28288/70269