﻿ $/Engineering/ReusableComponents/COMComponents/UCLIDRasterAndOCRMgmt/Core/Code/ $/Engineering/ProductDevelopment/Utils/SpatialStringViewer/Code
https://extract.atlassian.net/browse/ISSUE-11945
Added simple ruler on the left side that shows where page boundaries are with a line with the new page number under it.  The ruler also shows the page number of the first and last lines currently displayed on the screen.

https://extract.atlassian.net/browse/ISSUE-11943
Added a page navigation toolbar above the text that has buttons for first, last, next and previous pages as well as goto page box that also displays the current page in the format of <current page> of <total pages>

https://extract.atlassian.net/browse/ISSUE-11942
Added page confidence to the status displayed at the bottom has the text Page Con:

Reviewed by Steve (2/3)

William

[git-vault-id] Extract$/Engineering@28345/70334