﻿https://extract.atlassian.net/browse/ISSUE-12080 The Enhance OCR task and rule-object now requires a separate license.
https://extract.atlassian.net/browse/ISSUE-12087
While the Enhance OCR task will not be available or run without the Enhance OCR license, the rule object with “run” if it is not license, though it will not attempt to enhance any text; the uss file text will be used as-is. An application trace indicating that Enhance OCR is not licensed will be logged once per process in this case.

This is to allow the rule object to be used rulesets (particularly the NRS) without concern as to whether the processing machine will have Enhance OCR licensed. If it does, it will benefit from the rule object, otherwise the rules will run fine, it just won’t benefit from Enhance OCR.

https://extract.atlassian.net/browse/ISSUE-11937
Removed more IcoMaps license IDs from the licensing dat files.

Review pending…

[git-vault-id] Extract$/Engineering@28644/70643