﻿$/Engineering/ProductDevelopment/AFIntegrations/SharePoint/Redaction/Core/Code $/Engineering/ProductDevelopment/AFIntegrations/SharePoint/Redaction/Utilities/IDShieldForSPClient/Core/Code
https://extract.atlassian.net/browse/ISSUE-12024
https://extract.atlassian.net/browse/ISSUE-8079
Modified IDShieldForSPClient so that it would put the Verification window in front when clicking on either the “Redact Now” or “Verify” buttons

There is still a problem if an error occurs when opening the Verification UI, error dialogs could be displayed behind the active window. I will be creating an another SCR for this.

William

[git-vault-id] Extract$/Engineering@28669/70668