﻿$/Engineering/ProductDevelopment/AttributeFinder/IndustrySpecific/Redaction/Installation/IDShield $/Engineering/ProductDevelopment/AttributeFinder/Installation/UCLID FlexIndex SDK $/Engineering/ProductDevelopment/AttributeFinder/SilentInstalls $/Engineering/ProductDevelopment/DataEntry/LabDE/Installation/LabDE
https://extract.atlassian.net/browse/ISSUE-7540
Added .NET 3.5 and .NET 4.0 as prerequisites for LabDE, IDShield, and FlexIndex.  This means that .NET 3.5 and .NET 4.0 will be installed by these products on all OS’s.   However, Windows 8 and Windows 2012 will only install .NET if there is an internet connection, and if it fails the product will not be installed.
Also added checks in the silent installs to check for the Uninstall key for the products being installed after installation to verify that the install was successful.  If it was not it will display a message about which product did not install.

William

[git-vault-id] Extract$/Engineering@28920/70944