﻿$/Engineering/ProductDevelopment/AFIntegrations/SharePoint/Build $/Engineering/ProductDevelopment/AttributeFinder/Build $/Engineering/ProductDevelopment/Common $/Engineering/ProductDevelopment/DataEntry/LabDE/Build
https://extract.atlassian.net/browse/ISSUE-12100
Added changes needed in the batch and make files for a build.  I modified the files so that in the future changing the NUANCE_API_DIR in D:\Engineering\ProductDevelopment\Common\InitBuildEnv.bat will be all that is required when API is updated in the future.

Also created an environment variable for LEADTOOLS_API_DIR in the same batch file so that updating that API will only require the change in the one file.

William

[git-vault-id] Extract$/Engineering@28931/70955