﻿https://extract.atlassian.net/browse/ISSUE-12162 ImageFormatConverter no longer always outputs PDFs as bitonal; if the input document is color or grayscale, the output will be as well.
https://extract.atlassian.net/browse/ISSUE-12096
The Create Redacted Image task no longer always outputs PDFs as bitonal. If the source document is color or grayscale, the output will be as well.

This fix for this issue was complicated by a LeadTools bug that caused the degradation of color information when loading from color PDFs.

I opened ticket #2014040310001601 with LeadTools and they provided a patch from 11/29/1012, which fixes the issue. That patch is being checked in as part of this fix.

Reviewed by:
William (1/0)

[git-vault-id] Extract$/Engineering@28945/70969