﻿https://extract.atlassian.net/browse/ISSUE-12188 Fixed circumstance that likely led to the “Internal error” exception displayed on Troy’s VM for the Hennepin demo.
https://extract.atlassian.net/browse/ISSUE-6777
Fixed “Word highlight background operation aborted” applications from being periodically logged, particularly when saving/closing a document very quickly after switching to a new page with the word highlight/redaction tool active.

Finally, removed the need to have the DE framework briefly display each page of a document as a document is loaded in order to obtain the DPI of the images. This was a very heavy-handed way of getting page info and besides the extra events that interfered with debugging the above issues, it likely was impacting document load performance in DE verification.

Reviewed by:
William (0/0)

[git-vault-id] Extract$/Engineering@29187/71216